% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{adm_range_plot}
\alias{adm_range_plot}
\title{Plots the admissible range of flow variability}
\usage{
adm_range_plot(River_name = "Ebro", First_year = 1964, Last_year = 2011,
  Year_impact = 1988)
}
\arguments{
\item{River_name}{Name of the river as character (e.g.: River_name = "Ebro")}

\item{First_year}{First year to consider in the analysis starting on October 1st (e.g.: First_year = 1964)}

\item{Last_year}{First year to consider in the analysis finishing on September 30th (e.g.: Last_year = 2011)}

\item{Year_impact}{Year when the human impact started (the construction of a dam) (e.g.: Year_impact = 1988)}
}
\value{
Plots the admissible range of flow variability based on the flow data during the pre-impact period.
}
\description{
Plots the admissible range of flow variability
}
