% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{col_per_year}
\alias{col_per_year}
\title{Sorts the flow data per years - Each year is a column}
\usage{
col_per_year(First_year = 1964, Last_year = 2011)
}
\arguments{
\item{First_year}{First year to consider in the analysis starting on October 1st (e.g.: First_year = 1964)}

\item{Last_year}{Last year to consider in the analysis finishing on September 30th (e.g.: Last_year = 2011)}
}
\value{
The transformed dataframe per year is ready for calculations
}
\description{
Sorts the flow data per years - Each year is a column
}
\examples{
col_per_year(First_year=1964,Last_year=2011)
}
