% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{f_daily_costs6}
\alias{f_daily_costs6}
\title{f_daily_costs6 - Calculates the daily environmental costs of flow regulation}
\usage{
f_daily_costs6(First_day, Last_day, Year_evaluated, Year_impact, a_low, a_high,
  b_low, b_high)
}
\arguments{
\item{First_day}{First day of the year to consider in the analysis starting on October 1st written as character (e.g. "1964-10-01")}

\item{Last_day}{First day of the year to consider in the analysis finishing on September 30th written as character (e.g. "2011-09-30")}

\item{Year_evaluated}{Year when the environmental impact is evaluated}

\item{Year_impact}{Year when the human impact started (e.g. the construction of a dam)}

\item{a_low}{Coefficient a of Low-flow impact of function ku (e.g. 0.05)}

\item{a_high}{Coefficient a of High-flow impact of function ku (e.g. 0.01)}

\item{b_low}{Coefficient b of Low-flow impact of function ku (e.g. 2)}

\item{b_high}{Coefficient b of High-flow impact of function ku (e.g. 2)}
}
\value{
Calculates the daily environmental costs of flow regulation for a specific year evaluated.
}
\description{
f_daily_costs6 - Calculates the daily environmental costs of flow regulation
}
