% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{f_plot_adm_range4}
\alias{f_plot_adm_range4}
\title{f_plot_adm_range4 - Plots the admissible range of flow variability}
\usage{
f_plot_adm_range4(River_name, First_day, Last_day, Year_impact)
}
\arguments{
\item{River_name}{Name of the river as character (e.g. "Ebro")}

\item{First_day}{First day of the year to consider in the analysis starting on October 1st written as character (e.g. "1964-10-01")}

\item{Last_day}{First day of the year to consider in the analysis finishing on September 30th written as character (e.g. "2011-09-30")}

\item{Year_impact}{Year when the human impact started (e.g. the construction of a dam)}
}
\value{
Plots the admissible range of flow variability based on the flow data during the pre-impact period.
}
\description{
f_plot_adm_range4 - Plots the admissible range of flow variability
}
