% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{f_plot_impact5}
\alias{f_plot_impact5}
\title{f_plot_impact5 - Plots the daily environmental impact of flow regulation (high- and low-flow impact)}
\usage{
f_plot_impact5(River_name, First_day, Last_day, Year_evaluated, Year_impact)
}
\arguments{
\item{River_name}{Name of the river written as character (e.g. "Ebro")}

\item{First_day}{First day of the year to consider in the analysis starting on October 1st written as character (e.g. "1964-10-01")}

\item{Last_day}{First day of the year to consider in the analysis finishing on September 30th written as character (e.g. "2011-09-30")}

\item{Year_evaluated}{Year when the environmental impact is evaluated}

\item{Year_impact}{Year when the human impact started (e.g. the construction of a dam)}
}
\value{
Plots the daily environmental impact of flow regulation (high- and low-flow impact).
}
\description{
f_plot_impact5 - Plots the daily environmental impact of flow regulation (high- and low-flow impact)
}
