% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{f_summary_flow3}
\alias{f_summary_flow3}
\title{f_summary_flow3 - Provides a summary of flow data during the pre-impact period}
\usage{
f_summary_flow3(First_day, Last_day, Year_impact)
}
\arguments{
\item{First_day}{First day of the year to consider in the analysis starting on October 1st written as character (e.g. "1964-10-01")}

\item{Last_day}{First day of the year to consider in the analysis finishing on September 30th written as character (e.g. "2011-09-30")}

\item{Year_impact}{Year when the human impact started (e.g. the construction of a dam)}
}
\value{
Provides a dataframe on a daily basis of mean, min, p10, p25, median, p75, p90 and max values during the pre-impact period.
}
\description{
f_summary_flow3 - Provides a summary of flow data during the pre-impact period
}
