% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{f_years2}
\alias{f_years2}
\title{f_years2 - Sorts the flow data per years - Each year is a column}
\usage{
f_years2(First_day, Last_day)
}
\arguments{
\item{First_day}{First day of the year to consider in the analysis starting on October 1st written as character (e.g. "1964-10-01")}

\item{Last_day}{First day of the year to consider in the analysis finishing on September 30th written as character (e.g. "2011-09-30")}
}
\value{
The transformed dataframe per year is ready for calculations
}
\description{
f_years2 - Sorts the flow data per years - Each year is a column
}
