% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowdata}
\alias{flowdata}
\title{River water flow data.}
\format{A data frame with 17166 rows and 2 variables:
\describe{
\item{Date}{Date}
\item{Flow}{Water flow, in xxx}
...
}}
\source{
\url{http://doi...?}
}
\usage{
flowdata
}
\description{
A dataset containing daily river water flow data for a specific study area? from 01/10/1964 to 30/09/2011.
}
\keyword{datasets}
