% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{impact_reg_multi0}
\alias{impact_reg_multi0}
\title{Plots the daily environmental impact of flow regulation for multiple years}
\usage{
impact_reg_multi0(River_name = "Ebro", First_year = 1964,
  Last_year = 2011, Year_evaluated = 2010, Year_impact = 1988,
  x_coef = 0.8)
}
\arguments{
\item{River_name}{Name of the river written as character (e.g.: River_name = "Ebro")}

\item{First_year}{First year to consider in the analysis starting on October 1st (e.g.: First_year = 1964)}

\item{Last_year}{First year to consider in the analysis finishing on September 30th (e.g.: Last_year = 2011)}

\item{Year_evaluated}{Year when the environmental impact is evaluated (e.g.: Year_evaluated = 2010)}

\item{Year_impact}{Year when the human impact started (the construction of a dam) (e.g.: Year_impact = 1988)}

\item{x_coef}{A coeficient to change the font size in the graphs proportionally to the number of graphs plotted (e.g.: x_coef = 0.8)}
}
\value{
Plots the daily environmental impact of flow regulation for multiple years.
}
\description{
Plots the daily environmental impact of flow regulation for multiple years
}
