% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{structure_date}
\alias{structure_date}
\title{Transforms and reorders the flow data}
\usage{
structure_date(df = "", S_Day, S_Month, S_Year)
}
\arguments{
\item{df}{Data frame with river flow data}

\item{S_Day}{Position in Date string of the first digit of two-digits day}

\item{S_Month}{Position in Date string of the first digits of two-digits month}

\item{S_Year}{Position in Date string of the first digits of four-digits year}
}
\value{
The transformed dataframe on a daily basis sohuld now be ready for calculations
}
\description{
Transforms and reorders the flow data
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 structure_date(df='flowdata',S_Day=1,S_Month=4,S_Year=7)
 }
}
}
