% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{summary_flow}
\alias{summary_flow}
\title{Provides a summary of flow data during the pre-impact period}
\usage{
summary_flow(First_year = 1964, Last_year = 2011, Year_impact = 1988)
}
\arguments{
\item{First_year}{First year to consider in the analysis starting on October 1st (e.g.: First_year = 1964)}

\item{Last_year}{First year to consider in the analysis finishing on September 30th (e.g.: Last_year = 2011)}

\item{Year_impact}{Year when the human impact started (the construction of a dam) (e.g.: Year_impact = 1988)}
}
\value{
Provides a dataframe on a daily basis of mean, min, p10, p25, median, p75, p90 and max values during the pre-impact period.
}
\description{
Provides a summary of flow data during the pre-impact period
}
