% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{f_plot_adm_range4}
\alias{f_plot_adm_range4}
\title{Plots the admissible range of flow variability}
\usage{
f_plot_adm_range4(River_name, First_year, Last_year, Year_impact)
}
\arguments{
\item{River_name}{Name of the river as character (e.g.: River_name = "Ebro")}

\item{First_year}{First year to consider in the analysis starting on October 1st (e.g.: First_year = 1964)}

\item{Last_year}{First year to consider in the analysis finishing on September 30th (e.g.: Last_year = 2011)}

\item{Year_impact}{Year when the human impact started (the construction of a dam) (e.g.: Year_impact = 1988)}
}
\value{
Plots the admissible range of flow variability based on the flow data during the pre-impact period.
}
\description{
Plots the admissible range of flow variability
}
