% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowRegEnvCosts.R
\name{structure_date}
\alias{structure_date}
\title{Transforms and reorders the flow data}
\usage{
structure_date(dafra = "flowdata", S_Day = 1, S_Month = 4, S_Year = 7)
}
\arguments{
\item{dafra}{Data frame with river flow data (default value is empty)}

\item{S_Day}{Position in Date string of the first digit of two-digits day (default value is 1)}

\item{S_Month}{Position in Date string of the first digits of two-digits month (default value is 4)}

\item{S_Year}{Position in Date string of the first digits of four-digits year (default value is 7)}
}
\value{
A transformed data frame on a daily basis ready for further calculations
}
\description{
Transforms and reorders the flow data
}
\examples{
\dontrun{
data(flowdata)
structure_date(dafra='flowdata',S_Day=1,S_Month=4,S_Year=7)}
}
